server/client/requirements_client.txt
# HTTP requests
requests

# Image processing
Pillow
numpy
opencv-python

# Configuration and data
pyyaml

# Visualization
matplotlib

# Additional utilities
typing-extensions  # For better type hints

# Async HTTP
aiohttp